Bumped WIN32_IE version in blackbox to match the x64 headers. Commented out or restricted to 32 bit compilation some defintions in the blackbox header files. Fixed casts from pointers to user types the same size as pointers.
Also fixed stupid error I left in blackbox makefile

+++ blackbox/m_alloc.cpp
@@ -335,7 +335,7 @@ void operator delete[] (void *v)
 #else
 #ifdef __GNUC__  // get rid of bloated libstdc++ implementation
 #include <malloc.h>
-void * operator new (unsigned n)
+void * operator new (size_t n)
 {
 	return malloc(n);
 }
@@ -343,7 +343,7 @@ void operator delete (void *v)
 {
 	 free(v);
 }
-void * operator new[] (unsigned n)
+void * operator new[] (size_t n)
 {
 	return malloc(n);
 }
